requirements.txt
requests
line-bot-sdk
beautifulsoup4